api/requirements.txt
anyio==3.6.2
appdirs==1.4.4
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
cssselect==1.2.0
docker==6.0.1
fake-useragent==1.1.3
fastapi==0.95.0
idna==3.4
importlib-metadata==6.3.0
importlib-resources==5.12.0
lxml==4.9.2
packaging==23.0
parse==1.19.0
praw==7.7.0
prawcore==2.3.0
pydantic==1.10.7
pyee==8.2.2
pyppeteer==1.0.2
pyquery==2.0.0
requests==2.28.2
requests-html==0.10.0
sniffio==1.3.0
soupsieve==2.4
starlette==0.26.1
tqdm==4.65.0
typing_extensions==4.5.0
update-checker==0.18.0
urllib3==1.26.15
w3lib==2.1.1
websocket-client==1.5.1
websockets==10.4
youtube-transcript-api==0.5.0
zipp==3.15.0
